End the homepage products-browse experiment, remove the legacy template, and fix the search results z-index.
