+++ assets/scripts/main-dd-js.js
@@ -20,7 +20,6 @@ import './components/accordion-auto-open';
 import './components/product-section-collapse';
 import './components/signup';
 import './components/conversational-search';
-import './components/home-experiment';
 import './components/stepper';
 
 // Add Bootstrap Tooltip across docs
