+++ layouts/index.html
@@ -97,18 +97,7 @@
 
         </main>
 
-        {{- if eq .Site.Language.Lang "en" }}
-            <div class="home-content__variant-slot" data-variant-resolving>
-                <div data-home-variant-content="new" hidden>
-                    {{ partial "home/products-browse.html" . }}
-                </div>
-                <div data-home-variant-content="legacy" hidden>
-                    {{ partial "home/legacy.html" . }}
-                </div>
-            </div>
-        {{- else }}
-            {{ partial "home/products-browse.html" . }}
-        {{- end }}
+        {{ partial "home/products-browse.html" . }}
     </div>
 
     {{ partial "footer/footer.html" . }}
